+++ package.json
@@ -7,13 +7,13 @@
     "@ya.praktikum/react-developer-burger-ui-components": "^1.14.0",
     "clsx": "^1.2.1",
     "prop-types": "^15.8.1",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-dnd": "^14.0.5",
     "react-dnd-html5-backend": "^14.1.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-intersection-observer": "^8.32.5",
     "react-redux": "^7.2.9",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.1",
     "redux": "^4.2.1",
     "redux-devtools-extension": "^2.13.9",
