+++ src/pages/home/home.module.css
@@ -1,0 +1,12 @@
+.main {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  gap: 40px;
+}
+
+.loadingMessage {
+  display: flex;
+  align-items: center;
+  height: calc(100vh - 88px);
+}
