--- src/services/appContext.js
@@ -1,2 +1,0 @@
-import React from 'react';
-export const IngredientsContext = React.createContext();
