--- src/services/orderContext.js
@@ -1,2 +1,0 @@
-import React from 'react';
-export const OrderContext = React.createContext();
